% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_helpers.R
\name{.click_slider_id}
\alias{.click_slider_id}
\title{Which slider does a plot click move?}
\usage{
.click_slider_id(click_target)
}
\arguments{
\item{click_target}{The app's \code{input$click_target} value.}
}
\value{
\code{"iam"} if \code{click_target} is \code{"iam"}, otherwise \code{"i200"} (this is
also the safe default if \code{click_target} is missing or unrecognized).
}
\description{
Which slider does a plot click move?
}
\keyword{internal}
