% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_peakarea.R
\name{textile_amorphous_hump-methods}
\alias{textile_amorphous_hump-methods}
\alias{print.textile_amorphous_hump}
\alias{plot.textile_amorphous_hump}
\title{Methods for \code{textile_amorphous_hump} objects}
\usage{
\method{print}{textile_amorphous_hump}(x, ...)

\method{plot}{textile_amorphous_hump}(x, ...)
}
\arguments{
\item{x}{A \code{textile_amorphous_hump} from \code{\link[=fit_amorphous_hump]{fit_amorphous_hump()}}.}

\item{...}{Unused.}
}
\value{
\code{print()} and \code{plot()} return \code{x} invisibly (\code{plot()} after
printing a \code{ggplot}).
}
\description{
Methods for \code{textile_amorphous_hump} objects
}
\examples{
tt <- seq(10, 40, by = 0.05)
y <- 200 + 600 * exp(-((tt - 20.5) / 3)^2)
hump <- fit_amorphous_hump(tt, y, window = c(15, 28))
hump
plot(hump)
}
